pyproject.toml
[project]
name = "fine-tune-wikipedia"
version = "0.1.0"
description = ""
authors = [
    {name = "Pb mavericks",email = "patricia.brenny@mavs.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "datasets (>=4.0.0,<5.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "transformers (>=4.56.1,<5.0.0)"
]

[tool.poetry]
packages = [{include = "fine_tune_wikipedia", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
